Update LoginWindow.java
Made the LoginWindow the starting point of the application by moving the main method from MainInterface to LoginWindow.    

Added an action listener to the "Login" button to handle the authentication process. This listener retrieves the username and password entered by the user, calls the authenticateUser method to verify the credentials, and either opens the MainInterface (on success) or displays an error message (on failure). Currently just accepts any input.   

We added an action listener to the "Sign-Up" button to open the CreateAccountWindow and allow users to create a new account